[TEP-0075] Propose object (dicitonary) param and result types 🤓
This TEP proposes adding support for more structured params and results
by introducing objects (aka dictionaries), and introducing a (small)
subset of json schema syntax for declaring keys in the object.

This refers to TEP-0074 which will be added in a separate commit and
proposes deprecating PipelineResources.

This is based on @skaegi's proposals in https://github.com/tektoncd/pipeline/issues/1393